Move new test into existing InternalAuthenticationIntegrationTest
