+++ src/test/java/de/tum/in/www1/artemis/authentication/InternalAuthenticationIntegrationTest.java
@@ -10,6 +10,7 @@ import java.util.Set;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.authentication.TestingAuthenticationToken;
 import org.springframework.security.test.context.support.WithAnonymousUser;
@@ -20,8 +21,11 @@ import org.springframework.test.context.TestPropertySource;
 import de.tum.in.www1.artemis.domain.User;
 import de.tum.in.www1.artemis.repository.CourseRepository;
 import de.tum.in.www1.artemis.security.ArtemisInternalAuthenticationProvider;
+import de.tum.in.www1.artemis.security.jwt.TokenProvider;
 import de.tum.in.www1.artemis.service.UserService;
 import de.tum.in.www1.artemis.util.ModelFactory;
+import de.tum.in.www1.artemis.web.rest.UserJWTController;
+import de.tum.in.www1.artemis.web.rest.vm.LoginVM;
 import de.tum.in.www1.artemis.web.rest.vm.ManagedUserVM;
 
 @ActiveProfiles("artemis")
@@ -35,9 +39,14 @@ public class InternalAuthenticationIntegrationTest extends AuthenticationIntegra
     private ArtemisInternalAuthenticationProvider artemisInternalAuthenticationProvider;
 
     @Autowired
+    private TokenProvider tokenProvider;
+
+    @Autowired
     private CourseRepository courseRepository;
 
     private User student;
+
+    private static final String USERNAME = "student1";
 
     @Override
     @BeforeEach
@@ -45,7 +54,7 @@ public class InternalAuthenticationIntegrationTest extends AuthenticationIntegra
         database.addUsers(1, 0, 0);
         super.setUp();
 
-        student = userRepository.findOneWithGroupsAndAuthoritiesByLogin("student1").get();
+        student = userRepository.findOneWithGroupsAndAuthoritiesByLogin(USERNAME).get();
         final var encrPassword = userService.passwordEncoder().encode(USER_PASSWORD);
         student.setPassword(encrPassword);
         userRepository.save(student);
@@ -57,7 +66,7 @@ public class InternalAuthenticationIntegrationTest extends AuthenticationIntegra
     public void launchLtiRequest_authViaEmail_success() throws Exception {
         super.launchLtiRequest_authViaEmail_success();
 
-        final var updatedStudent = userRepository.findOneWithGroupsAndAuthoritiesByLogin("student1").get();
+        final var updatedStudent = userRepository.findOneWithGroupsAndAuthoritiesByLogin(USERNAME).get();
         assertThat(student).isEqualTo(updatedStudent);
     }
 
@@ -104,5 +113,21 @@ public class InternalAuthenticationIntegrationTest extends AuthenticationIntegra
         assertThat(response).isNotNull();
         assertThat(student).as("Returned user is equal to sent update").isEqualTo(response);
         assertThat(student).as("Updated user in DB is equal to sent update").isEqualTo(updatedUserIndDB);
+    }
+
+    @Test
+    @WithAnonymousUser
+    public void testJWTAuthentication() throws Exception {
+        LoginVM loginVM = new LoginVM();
+        loginVM.setUsername(USERNAME);
+        loginVM.setPassword(USER_PASSWORD);
+        loginVM.setRememberMe(true);
+
+        HttpHeaders httpHeaders = new HttpHeaders();
+        httpHeaders.add("User-Agent", "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/73.0.3683.103 Safari/537.36");
+
+        UserJWTController.JWTToken jwtToken = request.postWithResponseBody("/api/authenticate", loginVM, UserJWTController.JWTToken.class, HttpStatus.OK, httpHeaders);
+        assertThat(jwtToken.getIdToken()).as("JWT token is present").isNotNull();
+        assertThat(this.tokenProvider.validateToken(jwtToken.getIdToken())).as("JWT token is valid").isTrue();
     }
 }
